Palindrome
This solution is created using the Rust programming language.  The build is done using Cargo.

The solution uses regular expressions to remove all non letters from the string.

The string is then converted to lower case.  The next step is to take a copy and reverse the string.

The two strings are compared for equality and the result is printed to stdout.

The command line arguments are tested for options to display the usage.